@@ -1,39 +1,50 @@
 import * as React from "react";
-import {ICounter, ICounterAction} from "../models/counterModel";
-import {DECREMENT, INCREMENT} from "../redux/modules/counterModule";
-const {connect} = require("react-redux");
+import {connect} from "react-redux";
+import {IStore} from "../redux/IStore";
+import {IDispatchToProps} from "../redux/modules/baseModule";
+import {decrement, increment} from "../redux/modules/counterModule";
 
-interface IProps {
-  counter: ICounter;
-  increment: Redux.ActionCreator<ICounterAction>;
-  decrement: Redux.ActionCreator<ICounterAction>;
-}
+class CounterPage extends React.Component<IStateToProps & IDispatchToProps, null> {
+  constructor() {
+    super();
+    this.handleIncrement = this.handleIncrement.bind(this);
+    this.handleDecrement = this.handleDecrement.bind(this);
+  }
 
-@connect(
-  (state) => ({counter: state.counter}),
-  (dispatch) => ({
-    decrement: () => dispatch({type: DECREMENT}),
-    increment: () => dispatch({type: INCREMENT})
-  })
-)
+  public handleIncrement(): void {
+    this.props.dispatch(increment());
+  }
 
-class CounterPage extends React.Component<IProps, void> {
+  public handleDecrement(): void {
+    this.props.dispatch(decrement());
+  }
+
   public render(): JSX.Element {
-    const {increment, decrement, counter} = this.props;
+    const {count} = this.props;
 
     return (
       <div>
         <h4>Counter Example</h4>
-        <button name="incBtn" onClick={increment}>
+        <button name="incBtn" onClick={this.handleIncrement}>
           INCREMENT
         </button>
-        <button name="decBtn" onClick={decrement} disabled={counter.count <= 0}>
+        <button name="decBtn" onClick={this.handleDecrement} disabled={count <= 0}>
           DECREMENT
         </button>
-        <p>{counter.count}</p>
+        <p>{count}</p>
       </div>
     );
   }
 }
 
-export {CounterPage}
+interface IStateToProps {
+  count: number;
+}
+
+const mapStateToProps = (state: IStore) => ({
+  count: state.counter.payload.count
+});
+
+const connectedCounter = connect<IStateToProps, IDispatchToProps, null>(mapStateToProps)(CounterPage);
+
+export {CounterPage as UnconnectedCounter, connectedCounter as CounterPage}
