@@ -1,9 +1,12 @@
-import {renderComponent} from "../helpers/TestHelper";
+import {TestHelper} from "../helpers/TestHelper";
 import {Header, Styles} from "./Header";
 
 describe("<Header />", () => {
 
-  const component = renderComponent(Header);
+  const renderer = new TestHelper();
+  const component = renderer
+    .withTranslation({ languageData: { about: "About", counter: "Counter", stars: "Stars" }, locale: "en-GB" })
+    .mount(Header);
 
   it("Renders with correct style", () => {
     expect(component.find("nav")).toHaveClassName(Styles.nav);
