@@ -1,13 +1,13 @@
-import {shallow} from "enzyme";
-import * as React from "react";
+import {TestHelper} from "../helpers/TestHelper";
 import {App, Styles} from "./App";
 
 describe("<App />", () => {
-
-  const component = shallow(<App />);
+  const component = (new TestHelper())
+    .withState({language: {payload: {locale: "en", languageData: {}}}})
+    .mount(App);
 
   it("Renders with correct style", () => {
-    expect(component).toHaveClassName(Styles.container);
+    expect(component.find("section")).toHaveClassName(Styles.container);
   });
 
 });
