@@ -1,41 +1,85 @@
-import {renderComponent} from "../helpers/TestHelper";
-import {CounterPage} from "./CounterPage";
+import {shallow} from "enzyme";
+import * as React from "react";
+import {TestHelper} from "../helpers/TestHelper";
+import {IStore} from "../redux/IStore";
+import {IAction} from "../redux/modules/baseModule";
+import {DECREMENT, ICounter, INCREMENT} from "../redux/modules/counterModule";
+import {CounterPage, UnconnectedCounter} from "./CounterPage";
 
 /** Mock App. State */
-const state: object = {
-  counter: {count: 1}
+const state: Partial<IStore> = {
+  counter: {
+    isFetching: false,
+    payload: {
+      count: 1
+    }
+  }
 };
 
-describe("<CounterPage />", () => {
-
+describe("<Counter />", () => {
   let component;
 
   beforeEach(() => {
-    component = renderComponent(CounterPage, state);
+    component = (new TestHelper()).withState(state).mount(CounterPage);
   });
 
-  it("Renders header", () => {
+  it("renders header", () => {
     expect(component.find("h4")).toHaveText("Counter Example");
   });
 
-  it("Renders Increment and Decrement buttons", () => {
+  it("renders Increment and Decrement buttons", () => {
     expect(component.find("button")).toHaveLength(2);
   });
 
-  it("Renders counter value", () => {
+  it("renders counter value", () => {
     expect(component.find("p")).toHaveText("1");
   });
 
-  it("Calls the increment", () => {
-    expect(component.find({name: "incBtn"})).toBeDefined();
-    component.find({name: "incBtn"}).simulate("click");
-    expect(component.find("p")).toHaveText("2");
+  it("calls handleIncrement() when increment button is clicked", () => {
+    const spy = jest.spyOn(UnconnectedCounter.prototype, "handleIncrement");
+    const shallowComponent = shallow(<UnconnectedCounter dispatch={jest.fn()} count={0}/>);
+
+    expect(shallowComponent.find({name: "incBtn"})).toBeDefined();
+    expect(spy).not.toHaveBeenCalled();
+    shallowComponent.find({name: "incBtn"}).simulate("click");
+    expect(spy).toHaveBeenCalled();
   });
 
-  it("Calls the decrement", () => {
-    expect(component.find({name: "decBtn"})).toBeDefined();
-    component.find({name: "decBtn"}).simulate("click");
-    expect(component.find("p")).toHaveText("0");
+  it("calls handleDecrement() when decrement button is clicked", () => {
+    const spy = jest.spyOn(UnconnectedCounter.prototype, "handleDecrement");
+    const shallowComponent = shallow(<UnconnectedCounter dispatch={jest.fn()} count={0}/>);
+
+    expect(shallowComponent.find({name: "decBtn"})).toBeDefined();
+    expect(spy).not.toHaveBeenCalled();
+    shallowComponent.find({name: "decBtn"}).simulate("click");
+    expect(spy).toHaveBeenCalled();
   });
 
+  describe("handleIncrement()", () => {
+    it("dispatches INCREMENT action", () => {
+      const dispatch = jest.fn();
+      const shallowComponent = shallow(<UnconnectedCounter dispatch={dispatch} count={0}/>);
+      const expectedValue: IAction<ICounter> = {
+        type: INCREMENT
+      };
+
+      expect(dispatch).not.toHaveBeenCalled();
+      (shallowComponent as any).instance().handleIncrement();
+      expect(dispatch).toHaveBeenCalledWith(expectedValue);
+    });
+  });
+
+  describe("handleDecrement()", () => {
+    it("dispatches DECREMENT action", () => {
+      const dispatch = jest.fn();
+      const shallowComponent = shallow(<UnconnectedCounter dispatch={dispatch} count={0}/>);
+      const expectedValue: IAction<ICounter> = {
+        type: DECREMENT
+      };
+
+      expect(dispatch).not.toHaveBeenCalled();
+      (shallowComponent as any).instance().handleDecrement();
+      expect(dispatch).toHaveBeenCalledWith(expectedValue);
+    });
+  });
 });
