@@ -1,32 +1,55 @@
-import {renderComponent} from "../helpers/TestHelper";
-import {StarsPage} from "./StarsPage";
+import {shallow} from "enzyme";
+import * as React from "react";
+import {TestHelper} from "../helpers/TestHelper";
+import {IStore} from "../redux/IStore";
+import {IAction} from "../redux/modules/baseModule";
+import {IStars, LOAD_STARS} from "../redux/modules/starsModule";
+import {StarsPage, UnconnectedStars} from "./StarsPage";
 
 describe("<Stars />", () => {
-
-  it("Renders stars", () => {
-    const state: object = {
+  it("renders stars", () => {
+    const state: Partial<IStore> = {
       stars: {
+        isFetching: false,
         payload: {
           stargazers_count: 61
         }
       }
     };
 
-    const component = renderComponent(StarsPage, state);
-
+    const renderer = new TestHelper();
+    const component = renderer.withState(state).mount(StarsPage);
     expect(component.find("div")).toHaveText("61");
   });
 
-  it("Renders fetching", () => {
-    const state: object = {
+  it("renders fetching", () => {
+    const state: Partial<IStore> = {
       stars: {
-        payload: null
+        isFetching: true,
+        payload: {
+          stargazers_count: -1
+        }
       }
     };
 
-    const component = renderComponent(StarsPage, state);
-
+    const renderer = new TestHelper();
+    const component = renderer.withState(state).mount(StarsPage);
     expect(component.find("div")).toHaveText("Fetching Stars..");
   });
 
+  it("dispatches LOAD_STARS action before rendering if stargazers_count === -1", () => {
+    const dispatch = jest.fn();
+    const expectedValue: IAction<IStars> = {
+      type: LOAD_STARS
+    };
+    expect(dispatch).not.toHaveBeenCalled();
+    shallow(<UnconnectedStars dispatch={dispatch} isFetching={false} stargazers_count={-1}/>);
+    expect(dispatch).toHaveBeenCalledWith(expectedValue);
+  });
+
+  it("does not dispatch LOAD_STARS action before rendering if stargazers_count !== -1", () => {
+    const dispatch = jest.fn();
+    shallow(<UnconnectedStars dispatch={dispatch} isFetching={false} stargazers_count={10}/>);
+    expect(dispatch).not.toHaveBeenCalled();
+  });
 });
